# crAPIEncryption/api-specs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OWASP crAPI API",
    "version": "1.0"
  },
...
  "paths": {
...
    "/community/api/v2/community/posts/{postId}": {
      "get": {
        "operationId": "OP_COMMUNITY_GET_POST",
        "summary": "Get an specific post in the forum",
        "tags": [
          "Community / Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-levo-allowed-roles": [
          "ROLE_USER",
          "ROLE_MECHANIC",
          "ROLE_ADMIN"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": ""
          }
        }
      },
      "parameters": [
        {
          "in": "path",
          "name": "postId",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/community/api/v2/community/posts": {
      "post": {
        "operationId": "OP_COMMUNITY_CREATE_POST",
        "summary": "Create a new post in the forum",
        "tags": [
          "Community / Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-levo-allowed-roles": [
          "ROLE_USER",
          "ROLE_MECHANIC",
          "ROLE_ADMIN"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "CreatedAt",
                    "author",
                    "authorid",
                    "comments",
                    "content",
                    "id",
                    "title"
                  ],
                  "properties": {
                    "author": {
                      "type": "object",
                      "required": [
                        "created_at",
                        "email",
                        "nickname",
                        "profile_pic_url",
                        "vehicleid"
                      ],
                      "properties": {
                        "vehicleid": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "email": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string"
                        },
                        "profile_pic_url": {
                          "type": "string"
                        },
                        "nickname": {
                          "type": "string"
                        }
                      }
                    },
                    "id": {
                      "type": "string"
                    },
                    "authorid": {
                      "type": "number"
                    },
                    "content": {
                      "type": "string"
                    },
                    "CreatedAt": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "comments": {
                      "type": "array",
                      "items": {}
                    }
                  }
                }
              }
            },
            "description": ""
          }
        }
      },
      "parameters": []
    },
    "/community/api/v2/community/posts/{postId}/comment": {
      "post": {
        "operationId": "OP_COMMUNITY_POST_ADD_COMMENT",
        "summary": "Add a comment to an existing post in the forum",
        "tags": [
          "Community / Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-levo-allowed-roles": [
          "ROLE_USER",
          "ROLE_MECHANIC",
          "ROLE_ADMIN"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "CreatedAt",
                    "author",
                    "authorid",
                    "comments",
                    "content",
                    "id",
                    "title"
                  ],
                  "properties": {
                    "author": {
                      "type": "object",
                      "required": [
                        "created_at",
                        "email",
                        "nickname",
                        "profile_pic_url",
                        "vehicleid"
                      ],
                      "properties": {
                        "vehicleid": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "email": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string"
                        },
                        "profile_pic_url": {
                          "type": "string"
                        },
                        "nickname": {
                          "type": "string"
                        }
                      }
                    },
                    "id": {
                      "type": "string"
                    },
                    "authorid": {
                      "type": "number"
                    },
                    "content": {
                      "type": "string"
                    },
                    "CreatedAt": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "comments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "CreatedAt",
                          "author",
                          "content",
                          "id"
                        ],
                        "properties": {
                          "CreatedAt": {
                            "type": "string"
                          },
                          "author": {
                            "type": "object",
                            "required": [
                              "created_at",
                              "email",
                              "nickname",
                              "profile_pic_url",
                              "vehicleid"
                            ],
                            "properties": {
                              "vehicleid": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "email": {
                                "type": "string"
                              },
                              "created_at": {
                                "type": "string"
                              },
                              "profile_pic_url": {
                                "type": "string"
                              },
                              "nickname": {
                                "type": "string"
                              }
                            }
                          },
                          "content": {
                            "type": "string"
                          },
                          "id": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "description": ""
          }
        }
      },
      "parameters": [
        {
          "in": "path",
          "name": "postId",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/community/api/v2/community/posts/recent": {
      "get": {
        "summary": "Fetch the most recent posts in the forum.",
        "operationId": "OP_COMMUNITY_GET_RECENT_POSTS",
        "tags": [
          "Community / Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-levo-allowed-roles": [
          "ROLE_USER",
          "ROLE_MECHANIC",
          "ROLE_ADMIN"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Transfer-Encoding": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string",
                      "example": "chunked"
                    },
                    "example": "chunked"
                  }
                }
              },
              "Access-Control-Allow-Headers": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string",
                      "example": "Accept, Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization"
                    },
                    "example": "Accept, Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization"
                  }
                }
              },
              "Access-Control-Allow-Methods": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string",
                      "example": "POST, GET, OPTIONS, PUT, DELETE"
                    },
                    "example": "POST, GET, OPTIONS, PUT, DELETE"
                  }
                }
              },
              "Access-Control-Allow-Origin": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string",
                      "example": "*"
                    },
                    "example": "*"
                  }
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  },
                  "description": "Array of forum posts"
                },
                "example": [
                  {
                    "id": "ConZLXacq3MqhbLQDrbNLf",
                    "title": "Title 3",
                    "content": "Hello world 3",
                    "author": {
                      "nickname": "Hacker",
                      "email": "hacker@darkweb.com",
                      "vehicleid": "abac4018-5a38-466c-ab7f-361908afeab6",
                      "profile_pic_url": "",
                      "created_at": "2021-09-16T01:46:32.432Z"
                    },
                    "comments": [],
                    "authorid": 3,
                    "CreatedAt": "2021-09-16T01:46:32.432Z"
                  },
                  {
                    "id": "rj2md2VVDBjYUGNG6LmQ9e",
                    "title": "Title 2",
                    "content": "Hello world 2",
                    "author": {
                      "nickname": "Victim Two",
                      "email": "victim.two@example.com",
                      "vehicleid": "8b9edbde-d74d-4773-8c9f-adb65c6056fc",
                      "profile_pic_url": "",
                      "created_at": "2021-09-16T01:46:32.429Z"
                    },
                    "comments": [],
                    "authorid": 2,
                    "CreatedAt": "2021-09-16T01:46:32.429Z"
                  },
                  {
                    "id": "C68Hgjaow2jieF59LWzqTH",
                    "title": "Title 1",
                    "content": "Hello world 1",
                    "author": {
                      "nickname": "Victim One",
                      "email": "victim.one@example.com",
                      "vehicleid": "649acfac-10ea-43b3-907f-752e86eff2b6",
                      "profile_pic_url": "",
                      "created_at": "2021-09-16T01:46:32.413Z"
                    },
                    "comments": [],
                    "authorid": 1,
                    "CreatedAt": "2021-09-16T01:46:32.413Z"
                  }
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Post": {
        "title": "Post",
        "required": [
          "id",
          "title",
          "content",
          "author",
          "comments",
          "authorid",
          "CreatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "author": {
            "$ref": "#/components/schemas/Author"
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": ""
          },
          "authorid": {
            "type": "integer",
            "format": "int32"
          },
          "CreatedAt": {
            "type": "string"
          }
        },
        "example": {
          "id": "ConZLXacq3MqhbLQDrbNLf",
          "title": "Title 3",
          "content": "Hello world 3",
          "author": {
            "nickname": "Hacker",
            "email": "hacker@darkweb.com",
            "vehicleid": "abac4018-5a38-466c-ab7f-361908afeab6",
            "profile_pic_url": "",
            "created_at": "2021-09-16T01:46:32.432Z"
          },
          "comments": [],
          "authorid": 3,
          "CreatedAt": "2021-09-16T01:46:32.432Z"
        }
      },
      "Author": {
        "title": "Author",
        "required": [
          "nickname",
          "email",
          "vehicleid",
          "profile_pic_url",
          "created_at"
        ],
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "vehicleid": {
            "type": "string"
          },
          "profile_pic_url": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          }
        },
        "example": {
          "nickname": "Hacker",
          "email": "hacker@darkweb.com",
          "vehicleid": "abac4018-5a38-466c-ab7f-361908afeab6",
          "profile_pic_url": "",
          "created_at": "2021-09-16T01:46:32.432Z"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}